requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
jinja2==3.1.4
python-dotenv==1.0.1
playwright==1.48.0
aiosqlite==0.20.0
pandas==2.2.3
apscheduler>=3.10
